move routes and service building to app.rs
